+++ picodi/_picodi.py
@@ -9,7 +9,7 @@ from contextlib import asynccontextmanager, contextmanager
 from dataclasses import asdict, dataclass
 from typing import TYPE_CHECKING, Any, NamedTuple, ParamSpec, TypeVar, cast
 
-from picodi._internal import DummyAwaitable
+from picodi._internal import NullAwaitable
 from picodi._scopes import (
     GlobalScope,
     NullScope,
@@ -63,10 +63,8 @@ class InternalRegistry:
                     scope_class=(scope_class if override_scope else None),
                     # If the provider is already in use, keep it in use
                     # otherwise, use the new value. For example, if a provider
-                    # is already in use and we want to replace it with a resource,
-                    # we should keep it in use. If it's already a resource and we
-                    # want to replace it with a regular dependency, we should set
-                    # is_use to True.
+                    # is already in use and we want to replace it, `in_use=True`
+                    # should take precedence.
                     in_use=(provider.in_use or in_use),
                 )
                 if to_replace != provider:
@@ -146,8 +144,8 @@ class Registry:
     def clear(self) -> None:
         """
         Clear the registry. It will remove all dependencies and overrides.
-        This method will not close any resources. So you need to manually call
-        `shutdown_resources` before this method.
+        This method will not close any dependencies. So you need to manually call
+        `shutdown_dependencies` before this method.
         """
         with self._storage.lock:
             self._storage.deps.clear()
@@ -219,11 +217,11 @@ def inject(fn: Callable[P, T]) -> Callable[P, T]:
 
         @functools.wraps(fn)
         async def wrapper(*args: P.args, **kwargs: P.kwargs) -> T:
-            for scope in _scopes.values():
-                scope.enter_decorator()
-            bound, dep_arguments = _arguments_to_getters(
+            bound, dep_arguments, scopes = _arguments_to_getters(
                 args, kwargs, signature, is_async=True
             )
+            for scope in scopes:
+                scope.enter_decorator()
             for name, get_value in dep_arguments.items():
                 bound.arguments[name] = await get_value()
 
@@ -233,7 +231,7 @@ def inject(fn: Callable[P, T]) -> Callable[P, T]:
             else:
                 result = await result_or_gen  # type: ignore[misc]
 
-            for scope in _scopes.values():
+            for scope in scopes:
                 scope.exit_decorator()
                 coro = scope.close_local()
                 if coro is not None:
@@ -244,16 +242,16 @@ def inject(fn: Callable[P, T]) -> Callable[P, T]:
 
         @functools.wraps(fn)
         def wrapper(*args: P.args, **kwargs: P.kwargs) -> T:
-            for scope in _scopes.values():
-                scope.enter_decorator()
-            bound, dep_arguments = _arguments_to_getters(
+            bound, dep_arguments, scopes = _arguments_to_getters(
                 args, kwargs, signature, is_async=False
             )
+            for scope in scopes:
+                scope.enter_decorator()
             for name, get_value in dep_arguments.items():
                 bound.arguments[name] = get_value()
 
             result = fn(*bound.args, **bound.kwargs)
-            for scope in _scopes.values():
+            for scope in scopes:
                 scope.exit_decorator()
                 scope.close_local()
             return result
@@ -265,6 +263,7 @@ def dependency(*, scope_class: type[Scope] = NullScope) -> Callable[[TC], TC]:
     """
     Decorator to declare a dependency. You don't need to use it with default arguments,
     use it only if you want to change the scope of the dependency.
+    Should be placed last in the decorator chain (on top).
     """
 
     if scope_class not in _scopes:
@@ -279,41 +278,29 @@ def dependency(*, scope_class: type[Scope] = NullScope) -> Callable[[TC], TC]:
     return decorator
 
 
-def resource(fn: TC) -> TC:
+def init_dependencies() -> Awaitable:
     """
-    Decorator to declare a resource. Resource is a dependency that should be
-    called only once, cached and shared across the application.
-    On shutdown, all resources will be closed
-    (you need to call `shutdown_resources` manually).
-    Use it with a dependency generator function to declare a resource.
-    Should be placed last in the decorator chain (on top).
-    """
-    return dependency(scope_class=SingletonScope)(fn)
-
-
-def init_resources() -> Awaitable:
-    """
-    Call this function to close all resources. Usually, it should be called
+    Call this function to close dependencies. Usually, it should be called
     when your application is shutting down.
     """
-    async_resources = []
+    async_deps = []
     global_providers = _internal_registry.filter(
         lambda p: p.in_use and issubclass(p.scope_class, GlobalScope)
     )
     for provider in global_providers:
         if provider.is_async:
-            async_resources.append(_resolve_value_async(provider))
+            async_deps.append(_resolve_value_async(provider))
         else:
             _resolve_value(provider)
 
-    if async_resources:
-        return asyncio.gather(*async_resources)
-    return DummyAwaitable()
+    if async_deps:
+        return asyncio.gather(*async_deps)
+    return NullAwaitable()
 
 
-def shutdown_resources() -> Awaitable:
+def shutdown_dependencies() -> Awaitable:
     """
-    Call this function to close all resources. Usually, it should be called
+    Call this function to close dependencies. Usually, it should be called
     when your application is shut down.
     """
     tasks = []
@@ -397,7 +384,7 @@ class Provider:
 
 def _arguments_to_getters(
     args: P.args, kwargs: P.kwargs, signature: Signature, is_async: bool
-) -> tuple[BoundArguments, dict[str, Callable[[], Any]]]:
+) -> tuple[BoundArguments, dict[str, Callable[[], Any]], list[Scope]]:
     bound = signature.bind(*args, **kwargs)
     bound.apply_defaults()
     dependencies: dict[Provider, list[str]] = {}
@@ -408,12 +395,16 @@ def _arguments_to_getters(
     get_val = _resolve_value_async if is_async else _resolve_value
 
     dep_arguments = {}
+    scopes = []
     for provider, names in dependencies.items():
         get_value: Callable = functools.partial(get_val, provider)
         for name in names:
             dep_arguments[name] = get_value
+        scope = provider.get_scope()
+        if scope not in scopes:
+            scopes.append(scope)
 
-    return bound, dep_arguments
+    return bound, dep_arguments, scopes
 
 
 def _resolve_value(provider: Provider) -> Any: